DiscordBot/requirements.txt
discord.py
pygithub
python-dotenv